plug_rails_cookie_session_store v0.3.0 PlugRailsCookieSessionStore.MessageVerifier

MessageVerifier makes it easy to generate and verify messages which are signed to prevent tampering. For example, the cookie store uses this verifier to send data to the client. Although the data can be read by the client, he cannot tamper it.

Functions

sign(binary, secret)

Signs a binary according to the given secret.

verify(binary, secret)

Decodes and verifies the encoded binary was not tampared with.